+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - custom `changed` function is the same as using `Flags.Changed`, changed to use just the later
 - use `hydrated` parameter on "get time entry" endpoint instead of getting details individually
 - change in progress time entry using the current api
+- using "Hydrated" instead of "Full" to be consistent will the api
 
 ### Fixed